itpa/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
coreapi==2.3.3
Django==1.11.14
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-extra-fields==0.7
django-filter==1.0.2
django-rest-swagger==2.0.4
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djangorestframework-gis==0.13
djangorestframework-jsonapi==2.5.0
django-rest-auth==0.9.3
drfdocs==0.0.11
gunicorn==19.9.0
inflection==0.3.1
itypes==1.1.0
Markdown==2.6.11
openapi-codec==1.3.2
mysqlclient==1.3.13
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
